+++ crates/geng-window/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "geng-window"
+authors.workspace = true
+categories.workspace = true
+description.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu", "wasm32-unknown-unknown"]
+
+[dependencies]
+batbox-la.workspace = true
+anyhow.workspace = true
+ugli.workspace = true
+serde.workspace = true
+log.workspace = true
+gilrs.workspace = true
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+web-sys = { workspace = true, features = [
+    "MouseEvent",
+    "TouchEvent",
+    "TouchList",
+    "Touch",
+    "WheelEvent",
+    "KeyboardEvent",
+    "DomRect",
+    "CssStyleDeclaration",
+    "HtmlCanvasElement",
+] }
+wasm-bindgen.workspace = true
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+winit.workspace = true
+glutin.workspace = true
+glutin-winit.workspace = true
+raw-window-handle.workspace = true
+image.workspace = true
