+++ crates/ugli-raw/src/webgl/mod.rs
@@ -13,7 +13,9 @@ pub type SizeIPtr = i32;
 pub struct Context {
     inner: web_sys::WebGlRenderingContext,
     angle_instanced_arrays: web_sys::AngleInstancedArrays,
+    #[allow(dead_code)]
     oes_standard_derivatives: web_sys::OesStandardDerivatives,
+    #[allow(dead_code)]
     blend_minmax: web_sys::ExtBlendMinmax,
 }
